+++ source/toolkits/FileTk.h
@@ -1,0 +1,17 @@
+#ifndef TOOLKITS_FILETK_H_
+#define TOOLKITS_FILETK_H_
+
+#include "Common.h"
+
+class FileTk
+{
+	public:
+		static bool checkFileEmpty(std::string path);
+
+	private:
+		FileTk() {}
+};
+
+
+
+#endif /* TOOLKITS_FILETK_H_ */