+++ source/workers/RemoteWorker.cpp
@@ -1,3 +1,5 @@
+#include <chrono>
+#include <thread>
 #include <string>
 #include <sstream>
 #include "RemoteWorker.h"
@@ -398,13 +400,20 @@ void RemoteWorker::startBenchPhase()
 void RemoteWorker::waitForBenchPhaseCompletion(bool checkInterruption)
 {
 	bool firstRound = true;
-	size_t sleepUS = progArgs->getSvcUpdateIntervalMS() * 1000;
-	size_t firstRoundSleepUS = std::min( (size_t)500000, sleepUS/2); /* sleep half sec in first
-		round to have first result when live stats get printed for the first time after 1 sec */
+
+	size_t svcUpdateIntervalMS = progArgs->getSvcUpdateIntervalMS();
+	if(svcUpdateIntervalMS > (progArgs->getLiveStatsSleepMS() / 2) )
+		svcUpdateIntervalMS = progArgs->getLiveStatsSleepMS() / 2;
+
+	std::chrono::milliseconds sleepMS(svcUpdateIntervalMS );
+	std::chrono::milliseconds firstRoundSleepMS( std::min(500, (int)(sleepMS.count() / 2) ) ); /*
+		shorter first round sleep to have results when live stats get printed for the first time */
+	std::chrono::steady_clock::time_point lastSleepT = workersSharedData->phaseStartT;
 
 	while(numWorkersDone < progArgs->getNumThreads() )
 	{
-		usleep(firstRound ? firstRoundSleepUS : sleepUS);
+		lastSleepT += firstRound ? firstRoundSleepMS : sleepMS;
+		std::this_thread::sleep_until(lastSleepT);
 
 		if(checkInterruption)
 			checkInterruptionRequest();